+++ lib/castle/client.rb
@@ -66,6 +66,16 @@ module Castle
       @api.request(command)
     end
 
+    def impersonate(options = {})
+      options = Castle::Utils.deep_symbolize_keys(options || {})
+
+      return unless tracked?
+      options[:timestamp] ||= @timestamp if @timestamp
+
+      command = Castle::Commands::Impersonate.new(@context).build(options)
+      @api.request(command)
+    end
+
     def disable_tracking
       @do_not_track = true
     end